broke off socket handling logic into its own controller
